tests/requirements.txt
flask_mysqldb==0.2.0
pytest-flask==0.15.1
pytest-selenium==1.17.0
pytest==5.3.4
selenium==3.141.0
webcolors==1.11.1
